Add files via upload
Alien Invasion is a Python game where you get to shoot and destroy aliens. You control your spaceship using the spacebar to shoot and the arrow keys to move it left and right. At the start of the game, you have three ships. If an alien hits your ship or reaches the bottom of the screen, you lose one of your ships. This continues until you run out of all three ships, and that's when the game is over